==> src/2020/1_data_preproc.ipynb <==
# Pré-processamento de dados
Este arquivo faz o pré-tratamento dos dados que são obtidos do pelo caRtola (https://github.com/henriquepgomide/caRtola).

import numpy as np
import pandas as pd

É necessário definir a rodada atual na variável CURRENT_ROUND

CURRENT_ROUND = 38
